# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/bookings": {
      "get": {
        "operationId": "BookingsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all bookings",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get booking by ID",
        "tags": [
          "bookings"
        ]
      },
      "put": {
        "operationId": "BookingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update booking status/room",
        "tags": [
          "bookings"
        ]
      },
      "delete": {
        "operationId": "BookingsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel/Delete booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/complaints": {
      "get": {
        "operationId": "ComplaintsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all complaints",
        "tags": [
          "complaints"
        ]
      },
      "post": {
        "operationId": "ComplaintsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit new complaint",
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}": {
      "get": {
        "operationId": "ComplaintsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get complaint by ID",
        "tags": [
          "complaints"
        ]
      },
      "put": {
        "operationId": "ComplaintsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComplaintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update complaint status",
        "tags": [
          "complaints"
        ]
      },
      "delete": {
        "operationId": "ComplaintsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete complaint",
        "tags": [
          "complaints"
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all notifications",
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get notification by ID",
        "tags": [
          "notifications"
        ]
      },
      "put": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a notification",
        "tags": [
          "notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a notification",
        "tags": [
          "notifications"
        ]
      }
    }
  },
  "info": {
    "title": "PG Rental Management API",
    "description": "The PG Rental Management System using in-memory structures",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "property": {
            "type": "string"
          },
          "room": {
            "type": "string"
          },
          "checkIn": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "rent": {
            "type": "number"
          }
        },
        "required": [
          "tenant",
          "phone",
          "property",
          "room",
          "checkIn",
          "duration",
          "rent"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "room": {
            "type": "string"
          }
        }
      },
...
      "CreateComplaintDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string"
          },
          "property": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "property",
          "description"
        ]
      },
      "UpdateComplaintDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "resolution": {
            "type": "string"
          }
        }
      },
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "recipients": {
            "type": "number"
          },
          "by": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "message",
          "type",
          "priority",
          "recipients",
          "by"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}